+++ flutter_local_notifications_windows/lib/src/details/notification_to_xml.dart
@@ -3,44 +3,7 @@ import 'package:xml/xml.dart';
 import '../../flutter_local_notifications_windows.dart';
 import 'xml/details.dart';
 
-extension on DateTime {
-  String toIso8601StringTz() {
-    // Get offset
-    final Duration offset = timeZoneOffset;
-    final String sign = offset.isNegative ? '-' : '+';
-    final String hours = offset.inHours.abs().toString().padLeft(2, '0');
-    final String minutes =
-      offset.inMinutes.abs().remainder(60).toString().padLeft(2, '0');
-    final String offsetString = '$sign$hours:$minutes';
-    // Get first part of properly formatted ISO 8601 date
-    final String formattedDate = toIso8601String().split('.').first;
-    return '$formattedDate$offsetString';
-  }
-}
-
-extension on WindowsNotificationDetails {
-  /// XML attributes for the toast notification as a whole.
-  Map<String, String> get attributes => <String, String>{
-    if (duration != null) 'duration': duration!.name,
-    if (timestamp != null)
-      'displayTimestamp': timestamp!.toIso8601StringTz(),
-    if (scenario != null) 'scenario': scenario!.name,
-  };
-}
-
-/// Extensions on [WindowsProgressBar].
-extension ProgressBarXml on WindowsProgressBar {
-  /// The data bindings for this progress bar.
-  ///
-  /// To support dynamic updates, [buildXml] will inject placeholder strings
-  /// called data bindings instead of actual values. This can then be updated
-  /// dynamically later by calling
-  /// [FlutterLocalNotificationsWindows.updateProgressBar].
-  Map<String, String> get data => <String, String>{
-    '$id-progressValue': value?.toString() ?? 'indeterminate',
-    if (label != null) '$id-progressString': label!,
-  };
-}
+export 'xml/progress.dart';
 
 /// Converts a notification with [WindowsNotificationDetails] into XML.
 ///
