+++ flutter_local_notifications_windows/lib/src/plugin/base.dart
@@ -3,7 +3,7 @@ import 'package:meta/meta.dart';
 import 'package:timezone/timezone.dart';
 
 import '../details.dart';
-import '../details/notification_to_xml.dart';
+import '../details/xml/progress.dart';
 
 export 'package:flutter_local_notifications_platform_interface/flutter_local_notifications_platform_interface.dart';
 export 'package:timezone/timezone.dart';
